requirements.txt
streamlit==1.49.1
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
python-dotenv==1.1.1
openpyxl==3.1.5
ollama==0.5.4
numpy==2.3.3
requests==2.32.5